+++ CHANGELOG.md
@@ -1,9 +1,18 @@
 # C/C++ Config Change Log
 
-## Version 0.1.1: September 20, 2021
+## Version 0.2.0: September 25, 2021
+
+- **Improvement**: Added generation of **.clang-format** and **.editorconfig**
+- **Info**: If files are yet present, they won't be overridden
+
+## Version 0.1.2: September 22, 2021
+
+- **Bugfix**: Directory name fixed
+
+## Version 0.1.1: September 22, 2021
 
 - **Bugfix**: Filepath bugfix
 
-## Version 0.1.0: September 22, 2021
+## Version 0.1.0: September 20, 2021
 
 - **Info**: First release